fix(critical): restore Goal.Id to long type for production compatibility
- Goal.Id was changed from long to int in v2.8.0, causing OverflowException in production
- Production database contains old goal IDs as large timestamps (milliseconds since epoch)
- Restore Goal.Id, IGoalService, GoalService, FinanceService, Goals.razor to use long
- Update tests to use long literal (1L) for FindAsync calls
- All 41 tests passing

Closes: OverflowException at Microsoft.Data.Sqlite.SqliteDataReader.GetInt32

+++ FamilyFinance/Services/GoalService.cs
@@ -27,7 +27,7 @@ public class GoalService : IGoalService
             .ToListAsync();
     }
 
-    public async Task<Goal?> GetByIdAsync(int id)
+    public async Task<Goal?> GetByIdAsync(long id)
     {
         _logger.LogDebug("Fetching goal {GoalId}", id);
         return await _db.Goals.FirstOrDefaultAsync(g => g.Id == id && !g.IsDeleted);
@@ -81,7 +81,7 @@ public class GoalService : IGoalService
     // Legacy method for backward compatibility
     public async Task SaveAsync(Goal goal) => await SaveAsync(goal, null);
 
-    public async Task<ServiceResult> DeleteAsync(int id, string? userId = null)
+    public async Task<ServiceResult> DeleteAsync(long id, string? userId = null)
     {
         var goal = await _db.Goals.FirstOrDefaultAsync(g => g.Id == id);
         if (goal == null)
@@ -102,6 +102,6 @@ public class GoalService : IGoalService
     }
 
     // Legacy method for backward compatibility
-    public async Task DeleteAsync(int id) => await DeleteAsync(id, null);
+    public async Task DeleteAsync(long id) => await DeleteAsync(id, null);
 }
